+++ todoman/cli.py
@@ -8,6 +8,7 @@ from os.path import expanduser, isdir
 
 import click
 import click_log
+from click_default_group import DefaultGroup
 
 from todoman import exceptions, formatters
 from todoman.configuration import ConfigurationException, load_config
@@ -119,12 +120,6 @@ def _sort_callback(ctx, param, val):
 
 
 def validate_status(ctx=None, param=None, val=None):
-    # The default command doesn't run callbacks as expected, so it needs to
-    # specify the callback'd type. When `list` is called explicitly, this
-    # callback *IS* run, so we need to handle that edge case:
-    if not isinstance(val, str):
-        return val
-
     statuses = val.upper().split(',')
 
     if 'ANY' in statuses:
@@ -162,9 +157,16 @@ def _todo_property_options(command):
 
 class AppContext:
     def __init__(self):
-        self.config = None
         self.db = None
         self.formatter_class = None
+
+        self.init_config()
+
+    def init_config(self):
+        try:
+            self.config = load_config()
+        except ConfigurationException as e:
+            raise click.ClickException(e.args[0])
 
     @cached_property
     def ui_formatter(self):
@@ -191,7 +193,20 @@ _interactive_option = click.option(
     help='Go into interactive mode before saving the task.')
 
 
-@click.group(invoke_without_command=True)
+class ConfigurableDefaultGroup(DefaultGroup):
+
+    def get_command(self, ctx, cmd_name):
+        if cmd_name not in self.commands:
+            ctx = ctx.ensure_object(AppContext)
+            cmd_name = ctx.config['main']['default_command']
+        return super(DefaultGroup, self).get_command(ctx, cmd_name)
+
+
+@click.group(
+    cls=ConfigurableDefaultGroup,
+    default='default',  # click crashes in this is None
+    default_if_no_args=True,
+)
 @click_log.init('todoman')
 @click_log.simple_verbosity_option()
 @click.option('--colour', '--color', default=None,
@@ -209,10 +224,6 @@ _interactive_option = click.option(
 @catch_errors
 def cli(click_ctx, color, porcelain, humanize):
     ctx = click_ctx.ensure_object(AppContext)
-    try:
-        ctx.config = load_config()
-    except ConfigurationException as e:
-        raise click.ClickException(e.args[0])
 
     if porcelain and humanize:
         raise click.ClickException('--porcelain and --humanize cannot be used'
@@ -245,12 +256,6 @@ def cli(click_ctx, color, porcelain, humanize):
 
     # Make python actually use LC_TIME, or the user's locale settings
     locale.setlocale(locale.LC_TIME, "")
-
-    if not click_ctx.invoked_subcommand:
-        invoke_command(
-            click_ctx,
-            ctx.config['main']['default_command'],
-        )
 
 
 def invoke_command(click_ctx, command):
@@ -483,7 +488,7 @@ def move(ctx, list, ids):
               callback=_validate_startable_param, help='Show only todos which '
               'should can be started today (i.e.: start time is not in the '
               'future).')
-@click.option('--status', '-s', default=['NEEDS-ACTION', 'IN-PROCESS'],
+@click.option('--status', '-s', default='NEEDS-ACTION,IN-PROCESS',
               callback=validate_status, help='Show only todos with the '
               'provided comma-separated statuses. Valid statuses are '
               '"NEEDS-ACTION", "CANCELLED", "COMPLETED", "IN-PROCESS" or "ANY"'